@@ -1,52 +1,52 @@
 ext.versions = [
-        kotlin           : '1.4.30',
-        koin             : '2.2.2',
+        kotlin           : '1.5.30',
+        koin             : '2.2.3',
         retrofit         : '2.9.0',
-        okhttp           : '4.7.2',
-        coroutines       : '1.3.0',
+        okhttp           : '3.14.9',//4.7.2',
+        coroutines       : '1.4.2',
         lifecycle        : '2.3.0',
         navigation       : '2.3.3',
 
-        playcore         : '1.9.1',
+        playcore         : '1.8.2',
         playcoreKtx      : '1.8.1',
 
         javaVersion      : JavaVersion.VERSION_1_8,
         minSdkVersion    : 23,
-        targetSdkVersion : 30,
-        compileSdkVersion: 30,
+        targetSdkVersion : 31,
+        compileSdkVersion: 31,
         versionCode      : 118,
         versionName      : '1.2.6',
-        buildToolsVersion: '30.0.2',
+//        buildToolsVersion: '30.0.2',
 
-        detekt           : '1.0.1',
-        ktlint           : '0.40.0'
+        detekt           : '1.18.1',
+        ktlint           : '0.42.1'
 
 ]
 
 ext.deps = [
         kotlin    : [
-                jdk   : "org.jetbrains.kotlin:kotlin-stdlib:${versions.kotlin}",
+                jdk   : "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${versions.kotlin}",
                 plugin: "org.jetbrains.kotlin:kotlin-gradle-plugin:${versions.kotlin}"
         ],
 
         koin      : [
-                core     : "org.koin:koin-core:${versions.koin}",
-                android  : "org.koin:koin-android:${versions.koin}",
-                viewmodel: "org.koin:koin-android-viewmodel:${versions.koin}",
-                test     : "org.koin:koin-test:${versions.koin}",
+                core     : "io.insert-koin:koin-core:${versions.koin}",
+                android  : "io.insert-koin:koin-android:${versions.koin}",
+                viewmodel: "io.insert-koin:koin-androidx-viewmodel:${versions.koin}",
+                test     : "io.insert-koin:koin-test:${versions.koin}",
         ],
 
         'retrofit': [
                 core       : "com.squareup.retrofit2:retrofit:${versions.retrofit}",
                 converter  : "com.squareup.retrofit2:converter-moshi:${versions.retrofit}",
-                moshiKotlin: "com.squareup.moshi:moshi-kotlin:1.11.0"
+                moshiKotlin: "com.squareup.moshi:moshi-kotlin:1.12.0"
         ],
 
         google    : [
-                material: 'com.google.android.material:material:1.3.0',
+                material: 'com.google.android.material:material:1.4.0',
 
                 firebase: [
-                        bom        : 'com.google.firebase:firebase-bom:25.12.0',
+                        bom        : 'com.google.firebase:firebase-bom:28.3.1',
                         config     : 'com.google.firebase:firebase-config-ktx',
                         perf       : 'com.google.firebase:firebase-perf',
                         analytics  : 'com.google.firebase:firebase-analytics-ktx',
