@@ -7,7 +7,7 @@ import br.com.ricarlo.common.util.ViewState
 import com.noticiasnow.R
 import com.noticiasnow.databinding.ActivityDetailsNewsBinding
 import com.noticiasnow.model.NewsModel
-import org.koin.android.viewmodel.ext.android.viewModel
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class DetailsNewsActivity : BaseActivity<ActivityDetailsNewsBinding>() {
 
