@@ -10,7 +10,7 @@ import com.noticiasnow.R
 import com.noticiasnow.databinding.ActivitySplashBinding
 import com.noticiasnow.home.HomeActivity
 import com.noticiasnow.login.LoginActivity
-import org.koin.android.viewmodel.ext.android.viewModel
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class SplashActivity : BaseActivity<ActivitySplashBinding>() {
 
